requirements.txt
torch
torchvision
numpy
Pillow
streamlit
plotly